backend/pyproject.toml
[project]
name = "med-voice-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Vladyslav Chaliuk",email = "chaliukvladyslav@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12, <3.14"
dependencies = [
    "sqlalchemy (>=2.0.39,<3.0.0)",
    "fastapi[all] (>=0.115.12,<0.116.0)",
    "alembic (>=1.15.1,<2.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "pwdlib[argon2,bcrypt] (>=0.2.1,<0.3.0)",
    "fastapi-pagination (>=0.12.34,<0.13.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "makefun (>=1.15.6,<2.0.0)",
    "fastapi-filter (>=2.0.1,<3.0.0)",
    "aiofiles (>=24.1.0,<25.0.0)",
    "zeroconf (>=0.146.3,<0.147.0)"
]

[tool.poetry]
packages = [{include = "backend", from = "src"}]


[tool.poetry.group.test.dependencies]
pytest = "^8.3.5"
pytest-asyncio = "^0.26.0"


[tool.poetry.group.dev.dependencies]
ruff = "^0.11.2"


[tool.poetry.group.build.dependencies]
pyinstaller = "^6.12.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
